+++ Source/ansic/charset.h
@@ -1,4 +1,5 @@
 // charset.h
+// Revision 16-may-2003
 
 #ifndef CHARSET_H_
 #define CHARSET_H_
@@ -7,6 +8,7 @@ namespace charset {
 
 typedef unsigned char chardata [8];
 
+extern const chardata default_data [256];
 extern chardata data [256];
 
 }